requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
clockify==3.0.0
clockify-api-client==0.1.0
cycler==0.11.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==13.3.3
fonttools==4.31.2
idna==3.3
kiwisolver==1.4.0
matplotlib==3.5.1
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
